Refactor: Harvester, HQ to module
